Rename FieldBase folder
